Air BnB - All configuration related to Security : Authentication & Authorization part all coverd Here, Best way to refer the spring security config

+++ src/main/java/com/project/AirBNB/AirBNB/service/impl/HotelServiceImpl.java
@@ -8,11 +8,14 @@ import com.project.AirBNB.AirBNB.entity.Room;
 import com.project.AirBNB.AirBNB.exception.ResourceNotFoundException;
 import com.project.AirBNB.AirBNB.repository.HotelRepository;
 import com.project.AirBNB.AirBNB.repository.RoomRepository;
+import com.project.AirBNB.AirBNB.security.entity.User;
+import com.project.AirBNB.AirBNB.security.exception.UnAuthorisedException;
 import com.project.AirBNB.AirBNB.service.HotelService;
 import com.project.AirBNB.AirBNB.service.InventoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,6 +37,8 @@ public class HotelServiceImpl implements HotelService {
         log.info("Inside the create hotelData service impl layer");
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);  //First we map dto to entity
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel = hotelRepository.save(hotel); //Save it
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class); //Map back entity to Dto
@@ -54,6 +59,12 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -63,6 +74,10 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -76,6 +91,10 @@ public class HotelServiceImpl implements HotelService {
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -90,7 +109,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
         hotel.setActive(true);
 
         // assuming only do it once
